clk_register_clkdev(): handle callers needing format string
Many callers either use NULL or const strings for the third argument of
clk_register_clkdev.  For those that do not, this is a risk for format
strings being accidentally processed (for example in device names).  This
adds the missing "%s" arguments to make sure format strings will not leak
into the clkdev.

Signed-off-by: Kees Cook <keescook@chromium.org>
Cc: Michael Turquette <mturquette@baylibre.com>
Cc: Stephen Boyd <sboyd@codeaurora.org>
Cc: Russell King <linux@arm.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>