ocfs2: add feature document for online file check
This document describes the OCFS2 online file check feature.

OCFS2 is often used in high-availaibility systems.  However, OCFS2 usually
converts the filesystem to read-only on errors.  This may not be
necessary, since turning the filesystem read-only would affect other
running processes as well, decreasing availability.  Then, a mount option
(errors=continue) was introduced, which would return the EIO to the
calling process and terminate furhter processing so that the filesystem is
not corrupted further.  The filesystem is not converted to read-only, and
the problematic file's inode number is reported in the kernel log.  The
user can try to check/fix this file via online filecheck feature.

Signed-off-by: Gang He <ghe@suse.com>
Cc: Goldwyn Rodrigues <rgoldwyn@suse.de>
Cc: Mark Fasheh <mfasheh@suse.com>
Cc: Joel Becker <jlbec@evilplan.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>