mm: hugetlb: add VmHugetlbRSS: field in /proc/pid/status
Currently there's no easy way to get per-process usage of hugetlb pages,
which is inconvenient because applications which use hugetlb typically
want to control their processes on the basis of how much memory (including
hugetlb) they use.  So this patch simply provides easy access to the info
via /proc/pid/status.

This patch shouldn't change the OOM behavior (so hugetlb usage is ignored
as is now,) which I guess is fine until we have some strong reason to do
it.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Jörn Engel <joern@logfs.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Mike Kravetz <mike.kravetz@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>