scripts/kernel-doc: processing -nofunc for functions only
Docproc process EXPORT_SYMBOL(f1) macro and uses -nofunc f1 to avoid
duplicated documentation in the next call.  It works for most of the
cases, but there are some specific situations where a struct has the same
name of an already-exported function.

Current kernel-doc behavior ignores those structs and do not add them to
the final documentation.  This patch fixes it.

This is non-usual and the only case I've found is the drm_modeset_lock
(function and struct) defined in drm_modeset_lock.h and
drm_modeset_lock.c.  Considering this, it should only affect the DRM
documentation by including struct drm_modeset_lock to the final Docbook.

Signed-off-by: Danilo Cesar Lemes de Paula <danilo.cesar@collabora.co.uk>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Jonathan Corbet <corbet@lwn.net>
Cc: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>